Created chat.js and ChatController
