+++ Portal/Models/ChatViewModel.cs
@@ -2,7 +2,7 @@
 {
     public class ChatViewModel
     {
-        public string? Username { get; set; }
+        public string? ViewName { get; set; }
 
         [Required(ErrorMessage = "Message must be entered")]
         public string? Message { get; set; }
